+++ torchmetrics/classification/binned_precision_recall.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, List, Optional, Tuple, Union
+from warnings import warn
 
 import torch
 from torch import Tensor
@@ -52,8 +53,14 @@ class BinnedPrecisionRecallCurve(Metric):
 
     Args:
         num_classes: integer with number of classes. For binary, set to 1.
-        num_thresholds: number of bins used for computation. More bins will lead to more detailed
-            curve and accurate estimates, but will be slower and consume more memory. Default 100
+        num_thresholds: number of bins used for computation.
+
+            .. deprecated:: v0.4
+                Use `thresholds`. Will be removed in v0.5.
+
+        thresholds: list or tensor with specific thresholds or a number of bins from linear sampling.
+            It is used for computation will lead to more detailed curve and accurate estimates,
+            but will be slower and consume more memory.
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False. default: True
         dist_sync_on_step:
@@ -62,11 +69,15 @@ class BinnedPrecisionRecallCurve(Metric):
         process_group:
             Specify the process group on which synchronization is called. default: None (which selects the entire world)
 
+    Raises:
+        ValueError:
+            If ``thresholds`` is not a int, list or tensor
+
     Example (binary case):
         >>> from torchmetrics import BinnedPrecisionRecallCurve
         >>> pred = torch.tensor([0, 0.1, 0.8, 0.4])
         >>> target = torch.tensor([0, 1, 1, 0])
-        >>> pr_curve = BinnedPrecisionRecallCurve(num_classes=1, num_thresholds=5)
+        >>> pr_curve = BinnedPrecisionRecallCurve(num_classes=1, thresholds=5)
         >>> precision, recall, thresholds = pr_curve(pred, target)
         >>> precision
         tensor([0.5000, 0.5000, 1.0000, 1.0000, 1.0000, 1.0000])
@@ -81,7 +92,7 @@ class BinnedPrecisionRecallCurve(Metric):
         ...                      [0.05, 0.05, 0.75, 0.05, 0.05],
         ...                      [0.05, 0.05, 0.05, 0.75, 0.05]])
         >>> target = torch.tensor([0, 1, 3, 2])
-        >>> pr_curve = BinnedPrecisionRecallCurve(num_classes=5, num_thresholds=3)
+        >>> pr_curve = BinnedPrecisionRecallCurve(num_classes=5, thresholds=3)
         >>> precision, recall, thresholds = pr_curve(pred, target)
         >>> precision   # doctest: +NORMALIZE_WHITESPACE
         [tensor([0.2500, 1.0000, 1.0000, 1.0000]),
@@ -106,10 +117,11 @@ class BinnedPrecisionRecallCurve(Metric):
     def __init__(
         self,
         num_classes: int,
-        num_thresholds: int = 100,
+        thresholds: Optional[Union[Tensor, List[float]]] = None,
         compute_on_step: bool = True,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
+        num_thresholds: Optional[int] = 100,  # ToDo: remove in v0.5
     ):
         super().__init__(
             compute_on_step=compute_on_step,
@@ -118,14 +130,27 @@ class BinnedPrecisionRecallCurve(Metric):
         )
 
         self.num_classes = num_classes
-        self.num_thresholds = num_thresholds
-        thresholds = torch.linspace(0, 1.0, num_thresholds)
-        self.register_buffer("thresholds", thresholds)
+        if thresholds is None and num_thresholds is not None:
+            warn(
+                "Argument `num_thresholds` is deprecated in v0.4 and will be removed in v0.5."
+                " Use `thresholds` instead.", DeprecationWarning
+            )
+            thresholds = num_thresholds
+        if isinstance(thresholds, int):
+            self.num_thresholds = thresholds
+            thresholds = torch.linspace(0, 1.0, thresholds)
+            self.register_buffer("thresholds", thresholds)
+        elif thresholds is not None:
+            if not isinstance(thresholds, (list, Tensor)):
+                raise ValueError('Expected argument `thresholds` to either be an integer, list of floats or a tensor')
+            thresholds = torch.tensor(thresholds) if isinstance(thresholds, list) else thresholds
+            self.num_thresholds = thresholds.numel()
+            self.register_buffer("thresholds", thresholds)
 
         for name in ("TPs", "FPs", "FNs"):
             self.add_state(
                 name=name,
-                default=torch.zeros(num_classes, num_thresholds, dtype=torch.float32),
+                default=torch.zeros(num_classes, self.num_thresholds, dtype=torch.float32),
                 dist_reduce_fx="sum",
             )
 
@@ -185,12 +210,22 @@ class BinnedAveragePrecision(BinnedPrecisionRecallCurve):
     Args:
         num_classes: integer with number of classes. Not nessesary to provide
             for binary problems.
-        num_thresholds: number of bins used for computation. More bins will lead to more detailed
-            curve and accurate estimates, but will be slower and consume more memory. Default 100
+        num_thresholds: number of bins used for computation.
+
+            .. deprecated:: v0.4
+                Use `thresholds`. Will be removed in v0.5.
+
+        thresholds: list or tensor with specific thresholds or a number of bins from linear sampling.
+            It is used for computation will lead to more detailed curve and accurate estimates,
+            but will be slower and consume more memory
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False. default: True
         process_group:
             Specify the process group on which synchronization is called. default: None (which selects the entire world)
+
+    Raises:
+        ValueError:
+            If ``thresholds`` is not a list or tensor
 
     Example (binary case):
         >>> from torchmetrics import BinnedAveragePrecision
@@ -233,12 +268,22 @@ class BinnedRecallAtFixedPrecision(BinnedPrecisionRecallCurve):
     Args:
         num_classes: integer with number of classes. Provide 1 for for binary problems.
         min_precision: float value specifying minimum precision threshold.
-        num_thresholds: number of bins used for computation. More bins will lead to more detailed
-            curve and accurate estimates, but will be slower and consume more memory. Default 100
+        num_thresholds: number of bins used for computation.
+
+            .. deprecated:: v0.4
+                Use `thresholds`. Will be removed in v0.5.
+
+        thresholds: list or tensor with specific thresholds or a number of bins from linear sampling.
+            It is used for computation will lead to more detailed curve and accurate estimates,
+            but will be slower and consume more memory
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False. default: True
         process_group:
             Specify the process group on which synchronization is called. default: None (which selects the entire world)
+
+    Raises:
+        ValueError:
+            If ``thresholds`` is not a list or tensor
 
     Example (binary case):
         >>> from torchmetrics import BinnedRecallAtFixedPrecision
@@ -264,14 +309,16 @@ class BinnedRecallAtFixedPrecision(BinnedPrecisionRecallCurve):
         self,
         num_classes: int,
         min_precision: float,
-        num_thresholds: int = 100,
+        thresholds: Optional[Union[Tensor, List[float]]] = None,
         compute_on_step: bool = True,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
+        num_thresholds: int = 100,  # ToDo: remove in v0.5
     ):
         super().__init__(
             num_classes=num_classes,
             num_thresholds=num_thresholds,
+            thresholds=thresholds,
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
             process_group=process_group,